Test added for group maps.
